+++ sdk/core/azure-core/src/main/java/com/azure/core/http/rest/RequestOptions.java
@@ -11,7 +11,7 @@ import com.azure.core.util.BinaryData;
 import java.util.function.Consumer;
 
 /**
- * This class contains the options to customize a HTTP request. {@link RequestOptions} can be
+ * This class contains the options to customize an HTTP request. {@link RequestOptions} can be
  * used to configure the request headers, query params, the request body, or add a callback
  * to modify all aspects of the HTTP request.
  *
@@ -27,10 +27,16 @@ import java.util.function.Consumer;
  * </p>
  *
  * <p><strong>Creating an instance of RequestOptions</strong></p>
- * {@codesnippet com.azure.core.http.rest.requestoptions.instantiation}
+ * <!-- src_embed com.azure.core.http.rest.requestoptions.instantiation -->
+ * <pre>
+ * RequestOptions options = new RequestOptions&#40;&#41;
+ *     .setBody&#40;BinaryData.fromString&#40;&quot;&#123;&#92;&quot;name&#92;&quot;:&#92;&quot;Fluffy&#92;&quot;&#125;&quot;&#41;&#41;
+ *     .addHeader&#40;&quot;x-ms-pet-version&quot;, &quot;2021-06-01&quot;&#41;;
+ * </pre>
+ * <!-- end com.azure.core.http.rest.requestoptions.instantiation -->
  *
  * <p><strong>Configuring the request with JSON body and making a HTTP POST request</strong></p>
- * To <a href="https://petstore.swagger.io/#/pet/addPet">add a new pet to the pet store</a>, a HTTP POST call should
+ * To <a href="https://petstore.swagger.io/#/pet/addPet">add a new pet to the pet store</a>, an HTTP POST call should
  * be made to the service with the details of the pet that is to be added. The details of the pet are included as the
  * request body in JSON format.
  *
@@ -59,11 +65,50 @@ import java.util.function.Consumer;
  * To create a concrete request, Json builder provided in javax package is used here for demonstration. However, any
  * other Json building library can be used to achieve similar results.
  *
- * {@codesnippet com.azure.core.http.rest.requestoptions.createjsonrequest}
+ * <!-- src_embed com.azure.core.http.rest.requestoptions.createjsonrequest -->
+ * <pre>
+ * JsonArray photoUrls = Json.createArrayBuilder&#40;&#41;
+ *     .add&#40;&quot;https:&#47;&#47;imgur.com&#47;pet1&quot;&#41;
+ *     .add&#40;&quot;https:&#47;&#47;imgur.com&#47;pet2&quot;&#41;
+ *     .build&#40;&#41;;
+ *
+ * JsonArray tags = Json.createArrayBuilder&#40;&#41;
+ *     .add&#40;Json.createObjectBuilder&#40;&#41;
+ *         .add&#40;&quot;id&quot;, 0&#41;
+ *         .add&#40;&quot;name&quot;, &quot;Labrador&quot;&#41;
+ *         .build&#40;&#41;&#41;
+ *     .add&#40;Json.createObjectBuilder&#40;&#41;
+ *         .add&#40;&quot;id&quot;, 1&#41;
+ *         .add&#40;&quot;name&quot;, &quot;2021&quot;&#41;
+ *         .build&#40;&#41;&#41;
+ *     .build&#40;&#41;;
+ *
+ * JsonObject requestBody = Json.createObjectBuilder&#40;&#41;
+ *     .add&#40;&quot;id&quot;, 0&#41;
+ *     .add&#40;&quot;name&quot;, &quot;foo&quot;&#41;
+ *     .add&#40;&quot;status&quot;, &quot;available&quot;&#41;
+ *     .add&#40;&quot;category&quot;, Json.createObjectBuilder&#40;&#41;.add&#40;&quot;id&quot;, 0&#41;.add&#40;&quot;name&quot;, &quot;dog&quot;&#41;&#41;
+ *     .add&#40;&quot;photoUrls&quot;, photoUrls&#41;
+ *     .add&#40;&quot;tags&quot;, tags&#41;
+ *     .build&#40;&#41;;
+ *
+ * String requestBodyStr = requestBody.toString&#40;&#41;;
+ * </pre>
+ * <!-- end com.azure.core.http.rest.requestoptions.createjsonrequest -->
  *
  * Now, this string representation of the JSON request can be set as body of RequestOptions
  *
- * {@codesnippet com.azure.core.http.rest.requestoptions.postrequest}
+ * <!-- src_embed com.azure.core.http.rest.requestoptions.postrequest -->
+ * <pre>
+ * RequestOptions options = new RequestOptions&#40;&#41;
+ *     .addRequestCallback&#40;request -&gt; request
+ *         &#47;&#47; may already be set if request is created from a client
+ *         .setUrl&#40;&quot;https:&#47;&#47;petstore.example.com&#47;pet&quot;&#41;
+ *         .setHttpMethod&#40;HttpMethod.POST&#41;
+ *         .setBody&#40;requestBodyStr&#41;
+ *         .setHeader&#40;&quot;Content-Type&quot;, &quot;application&#47;json&quot;&#41;&#41;;
+ * </pre>
+ * <!-- end com.azure.core.http.rest.requestoptions.postrequest -->
  */
 final class RequestOptions {
     private Consumer<HttpRequest> requestCallback = request -> { };
@@ -78,7 +123,7 @@ final class RequestOptions {
     }
 
     /**
-     * Gets whether or not to throw an exception when an HTTP response with a status code indicating an error
+     * Gets whether an exception is thrown when an HTTP response with a status code indicating an error
      * (400 or above) is received.
      *
      * @return true if to throw on status codes of 400 or above, false if not. Default is true.
@@ -125,7 +170,7 @@ final class RequestOptions {
      *
      * @param parameterName the name of the query parameter
      * @param value the value of the query parameter
-     * @param encoded whether or not this query parameter is already encoded
+     * @param encoded whether this query parameter is already encoded
      * @return the modified RequestOptions object
      */
     public RequestOptions addQueryParam(String parameterName, String value, boolean encoded) {
@@ -163,8 +208,8 @@ final class RequestOptions {
     }
 
     /**
-     * Sets whether or not to throw an exception when an HTTP response with a status code indicating an error
-     * (400 or above) is received. By default an exception will be thrown when an error response is received.
+     * Sets whether exception is thrown when an HTTP response with a status code indicating an error
+     * (400 or above) is received. By default, an exception will be thrown when an error response is received.
      *
      * @param throwOnError true if to throw on status codes of 400 or above, false if not. Default is true.
      * @return the modified RequestOptions object